Add utility class for validation
